+++ test/lit.cfg
@@ -342,7 +342,7 @@ shutil.rmtree(completion_cache_path, ignore_errors=True)
 ccp_opt = "-completion-cache-path %r" % completion_cache_path
 lit_config.note("Using code completion cache: " + completion_cache_path)
 
-config.substitutions.append( ('%incparse-test', '%utils/incrparse_test_util.py --temp-dir %t --swift-syntax-test %swift-syntax-test') )
+config.substitutions.append( ('%validate-incrparse', '%utils/incrparse/validate_parse.py --temp-dir %t --swift-syntax-test %swift-syntax-test') )
 config.substitutions.append( ('%swift_obj_root', config.swift_obj_root) )
 config.substitutions.append( ('%swift_src_root', config.swift_src_root) )
 config.substitutions.append( ('%{python}', sys.executable) )
